+++ .github/CHANGELOG.md
@@ -52,6 +52,7 @@
 1. [SD] Connect SD BLEED page APU bleed pressure indication to correct ADRs - @BlueberryKing (BlueberryKing)
 1. [FMS] Use station declination when appropriate for fix info and place/bearing radials - @tracernz (Mike)
 1. [MCDU] Fixed the FMGC annunciator light not illuminating - @tracernz (Mike)
+1. [COND] Add Air Conditioning systems failures - @mjuhe (Miquel Juhe)
 1. [EFB] Random Failures - @Maverickwoe (Garoomf)
 
 ## 0.10.0